+++ web/src/main/java/by/carlab/restControllers/CarDeleteRestController.java
@@ -1,6 +1,7 @@
 package by.carlab.restControllers;
 
 import by.carlab.cars.CarService;
+import by.carlab.orders.OrderService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.annotation.Secured;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,11 +14,16 @@ public class CarDeleteRestController {
     @Autowired
     private CarService carService;
 
-    //curl -s http://localhost:8080/rent/rest/admin/delete/car/i.html --user username:password
+    @Autowired
+    private OrderService orderService;
+
+    //curl -s http://localhost:8080/rent/rest/admin/delete/car/i.html --user admin:admin
 
     @GetMapping("/rest/admin/delete/car/{id}.html")
     @Secured(value = {"ROLE_ADMIN"})
     public void deleteCar(@PathVariable("id") int id) {
-        carService.deleteCar(id);
+        if(!orderService.isCarInOrder(id)) {
+            carService.deleteCar(id);
+        }
     }
 }
